LF-4158 move ImagePicker component to /component
