+++ packages/webapp/src/components/ImagePicker/styles.module.scss
